moved s3 tools to commands/__init__.py
