+++ bakery/management/commands/unpublish.py
@@ -1,7 +1,7 @@
 import logging
 from django.conf import settings
 from django.core.management.base import CommandError
-from bakery.management.commands.base_publish import BasePublishCommand
+from bakery.management.commands import BasePublishCommand
 
 logger = logging.getLogger(__name__)
